requirements.txt
attrs==21.4.0
decorator==5.1.1
docopt==0.6.2
h5py==3.6.0
iniconfig==1.1.1
mango-agents==0.1.2
mosaik==3.0.0
mosaik-api==3.0.0
mosaik-hdf5==0.4
networkx==2.5
numpy==1.22.0
packaging==21.3
paho-mqtt==1.5.1
pluggy==1.0.0
protobuf==3.13.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-asyncio==0.17.0
simpy==3.0.13
simpy.io==0.2.3
six==1.16.0
toml==0.10.2